requirements.txt
absl-py==0.7.1
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==1.5.0
bottle==0.12.16
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
coverage==4.5.3
cryptography==2.7
Cython==0.29.7
decorator==4.4.0
defusedxml==0.6.0
Django==2.1.4
django-appconf==1.0.3
django-autoslug==1.9.4
django-compressor==2.3
django-debug-toolbar==1.11
django-haystack==2.8.1
django-ipware==2.1.0
django-mdeditor==0.1.14
django-uuslug==1.1.8
djangorestframework==3.9.4
dwebsocket==0.5.10
easydict==1.9
elasticsearch==7.0.2
elasticsearch-dsl==7.0.0
entrypoints==0.3
gast==0.2.2
grpcio==1.21.1
h5py==2.7.0
html5lib==0.9999999
idna==2.8
ipaddress==1.0.22
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
jedi==0.13.3
jieba==0.39
Jinja2==2.10.1
jsonpickle==1.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.1.5
lazy-object-proxy==1.4.1
lxml==4.3.4
mahotas==1.4.5
Markdown==3.1.1
markdown2==2.3.8
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==1.4.2.post1
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
olefile==0.46
opencv-contrib-python==4.0.0.21
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
pycparser==2.19
Pygments==2.4.2
pylint==2.3.1
PyMySQL==0.9.3
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
python-logstash==0.4.6
python-memcached==1.59
python-slugify==3.0.2
pytz==2019.1
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.4
raven==6.10.0
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
torch==0.4.1
torchvision==0.2.2.post3
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.4.0
Unidecode==1.1.0
urllib3==1.25.3
wcwidth==0.1.7
web.py==0.40.dev0
webencodings==0.5.1
Werkzeug==0.15.4
WeRoBot==1.9.0
Whoosh==2.7.4
widgetsnbextension==3.4.2
wincertstore==0.2
wrapt==1.11.1
xlrd==1.2.0
xmltodict==0.12.0
